% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{autoquote.character}
\alias{autoquote.character}
\title{Autoquote Character}
\usage{
\method{autoquote}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Autoquote character.  Quotes elements with commas or double quotes; the latter are doubled.
}
\seealso{
Other autoquote: \code{\link{autoquote.default}},
  \code{\link{autoquote.factor}}, \code{\link{autoquote}}
}
\keyword{internal}
