% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{autoquote.default}
\alias{autoquote.default}
\title{Autoquote Default}
\usage{
\method{autoquote}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Autoquote default.  Returns its argument.
}
\seealso{
Other autoquote: \code{\link{autoquote.character}},
  \code{\link{autoquote.factor}}, \code{\link{autoquote}}
}
\keyword{internal}
