% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{autoquote.factor}
\alias{autoquote.factor}
\title{Autoquote Factor}
\usage{
\method{autoquote}{factor}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Autoquote character.  Quotes levels with commas or double quotes; the latter are doubled.
}
\seealso{
Other autoquote: 
\code{\link{autoquote.character}()},
\code{\link{autoquote.default}()},
\code{\link{autoquote}()}
}
\concept{autoquote}
\keyword{internal}
