% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{autoquote}
\alias{autoquote}
\title{Autoquote}
\usage{
autoquote(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Autoquote.  Generic, with default, character, and factor methods.
}
\seealso{
Other autoquote: 
\code{\link{autoquote.character}()},
\code{\link{autoquote.default}()},
\code{\link{autoquote.factor}()}
}
\concept{autoquote}
\keyword{internal}
